finishing up backend

+++ backend/controllers/product.controller.js
@@ -35,9 +35,7 @@ export const updateProduct = async (req, res) => {
   const { id } = req.params;
   const product = req.body;
   if (!mongoose.Types.ObjectId.isValid(id)) {
-    return res
-      .status(404)
-      .json({ success: false, message: "Invalid product id" });
+    return res.status(404).json({ success: false, message: "Invalid product id" });
   }
   try {
     const updatedProduct = await Product.findByIdAndUpdate(id, product, {
@@ -51,6 +49,9 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({ success: false, message: "Invalid product id" });
+  }
   // console.log('id:',id);
   try {
     await Product.findByIdAndDelete(id);
@@ -59,7 +60,7 @@ export const deleteProduct = async (req, res) => {
       .json({ success: true, message: "Product deleted successfully" });
   } catch (error) {
     console.log("Error in Delete product:", error.message);
-    res.status(404).json({ success: false, message: "Product not found" });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
